feat: add openai api
